Move proposed app instance info
